pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "resume"
version = "0.1.0"
classifiers = [
    "Private :: Do Not Upload",
]
requires-python = ">= 3.13"
dependencies = [
    "pydantic>=2.10.2",
    "jinja2>=3.1.4",
    "typer>=0.14.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
]

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit>=4.0.1",
    "ruff~=0.8.1",
]

[tool.rye.scripts]
resume = { call = "resume.app" }

setup = { chain = ["setup:sync", "setup:pre-commit"] }
"setup:sync" = "rye sync"
"setup:pre-commit" = "pre-commit install"

[tool.hatch.build.targets.wheel]
packages = [
    "src/resume",
]

# linting/formatting

[tool.ruff]
extend-exclude = [
    "**/venv",
    "**/.*",
    "**/__pycache__",
]

[tool.ruff.lint]
extend-select = [
    "I", # import sorting
    "RUF022", # __all__ sorting
]
extend-ignore = [
    # covered by Pyright
    "F821", # undefined variable
    "F401", # imported but unused
    "F841", # unused variable
]

[tool.ruff.lint.isort]
combine-as-imports = true

# type checking

[tool.pyright]
venvPath = "."
venv = ".venv"

exclude = [
    "venv",
    ".venv",
]

typeCheckingMode = "basic"

enableExperimentalFeatures = true
strictDictionaryInference = true
strictListInference = true
strictSetInference = true

reportAssertAlwaysTrue = "error"
reportConstantRedefinition = "error"
reportDeprecated = "error"
reportFunctionMemberAccess = "error"
reportIncompatibleMethodOverride = "error"
reportIncompatibleVariableOverride = "error"
reportIncompleteStub = "error"
reportInconsistentConstructor = "error"
reportInvalidStringEscapeSequence = "error"
reportInvalidStubStatement = "error"
reportInvalidTypeVarUse = "error"
reportMatchNotExhaustive = "error"
reportMissingParameterType = "error"
reportOverlappingOverload = "error"
reportSelfClsParameterName = "error"
reportTypeCommentUsage = "error"
reportUnknownParameterType = "error"
reportUnnecessaryCast = "error"
reportUnnecessaryContains = "error"
reportUnsupportedDunderAll = "error"
reportUntypedBaseClass = "error"
reportUntypedClassDecorator = "error"
reportUntypedFunctionDecorator = "error"
reportUntypedNamedTuple = "error"
reportWildcardImportFromLibrary = "error"

reportMissingTypeArgument = "warning"
reportPrivateUsage = "warning"
reportUnknownArgumentType = "warning"
reportUnknownLambdaType = "warning"
reportUnknownMemberType = "warning"
reportUnknownVariableType = "warning"
reportUnnecessaryComparison = "warning"
reportUnnecessaryIsInstance = "warning"
reportUnusedClass = "warning"
reportUnusedExpression = "warning"
reportUnusedFunction = "warning"
reportUnusedImport = "warning"
reportUnusedVariable = "warning"

reportMissingTypeStubs = "none"
reportDuplicateImport = "none"
